# ENH, SIMD: improve argmax/argmin performance