@mattip 2022-01-19
Thanks, unrolling by 4 was the thing that caught my eye, and your explanation makes it clear.