@seiko2plus 2022-01-20
@mattip,
> except bool take approximately the same time

for two reasons:
1. `BOOL_argmin()` isn't included by pr because `memchr` is well optimized by GCC and CLANG.

https://github.com/numpy/numpy/blob/ed84326d78d31e9b989ed59e541854269aa08812/numpy/core/src/multiarray/arraytypes.c.src#L3262-L3274

2. `BOOL_argmax()` was already vectorized by raw SIMD for both x86 and aarch64. note on VSX it shows pretty good improvement:
```Bash
-       214±0.5μs       28.0±0.1μs     0.13  bench_reduce.ArgMax.time_argmax(<class 'bool'>)
```

> that there is a cost to processing nans.

that's true but I can give the priority for numeric values and gain around 40% speed but that will cause a pretty bad impact on nanargmax/nanargmin performance. my idea is to invert the FP test and avoid breaking the loop. By the way, the nan benchmark isn't fear enough, and it doesn't expose the real improvements. see:
https://github.com/numpy/numpy/blob/ed84326d78d31e9b989ed59e541854269aa08812/benchmarks/benchmarks/bench_lib.py#L72-L74