@seberg 2022-02-13
We have a "performance" category.  So we do, but are very very spotty about it.  It is also OK to add multiple for each I think, just make sure to use the brief bullet format `* Improved performance of ...`.